PDP11: Reworked PCLK device to minimize scheduled events
Rather than use a simh event to increment/decrement the counter register
for each 'clock cycle', compute the time the interrupt will eventually fire
and schedule a single event to deal with that.  If the clock counter is
referenced before the interrupt fires, interpolate the value based on the
number of instructions that have actually been executed.

sim_activate_after() is used here, and the time value for the scheduled
activation is a uint32 representing microseconds.  The range of values
that the PCLK can programmatically specify could be large enough
to exceed an uint32 when number of microseconds is converted to
a count of instructions.  This is an issue that has minimal consequences,
but should be addressed internal to the implementation of
sim_activate_after().