Visual Studio Projects: Make sure ROM checks used most recent BuildROMs
The BuildROMs project now runs the BuildROMs program that was most
recently built (Debug or Release).  This will identify ROM data inconsistencies
that may exist at build time.

When building other projects which may depend on ROM data, the BuildROMs
is run as a sanity check and if it fails we make sure that we hadn't run an out
of date Release build version of the BuildROMs binary.