VAX780, VAX8600: Fixed bad test for UBA intr level (COVERITY)
IPL_UBA already has the subtract built in:

So it shouldn't be extracted again.

The whole routine looks a little strange, but the way it works is
that an interrupt from the UBA itself sets <bit 31> in the returned vector.
Because the vector is read by code and not used by hardware, the flag
bit is "harmless."

UBA interrupts occur only under strange circumstances, like bad map
pages and device NXMs. Under the simulator, with a debugged OS, they
never happen.