PDP8: Annotate fall through paths in switch statements
The code is a bit difficult to understand, but it represents the 'normal'
path for processing a DECtape word. The code always flows all the
way to the break.

For start of block, there's extra code to check for a timing error; but
then the first word is processed (case 0).

For a normal word, a 3-cycle data break is done - increment word
count and current address, check for word count overflow, put the
word in the buffer; but then check for end of block (case DTO_WCO).

If the word count has already overflowed,  just check for end of block.

So yes, the end of case 0 should be labeled 'fall through' as well.

FNC_WRIT has the same structure and needs the same comment.