Visual Studio Projects: Properly populate .git-commit-id after git clone
This only happens if git is installed locally and available in the current
path when Visual Studio executes.  Whether git is in the path is an
installation option when git for Windows is installed.

Otherwise, proper results should be available after subsequent
activity on the local repository.  Any build will install the git hooks
that properly populate .git-commit-id when any changes are made
to the local repo.