I1620: Fixed bug in select index A (COVERITY)
If the CPU is a model 1, then the value of idxb is irrelevant. For
tidiness, it should always be 0, but on a model 1, idxe (index
enable) is always 0, so idxb is not looked at. Thus,

- on a model 1, idxb is set to 0, which is harmless and probably a good
   thing to do.
- on a model 2, idxe is set to 1, and idxb is set to 0, which is the correct
   behavior.

Note that case 0 doesn't need to check for model 2, because clearing
idxe and idxb yields the correct settings on a model 1: indexing is off.