Visual Studio Projects: Assure .git-commit-id.h gets updated correctly
Some git capable environments on Windows don't honor the installed git
hook scripts in the repository.  Now that, if we're working within a git repo,
we force a working version of git to be installed in the runtime path.  We
leverage that to force a validation of the commit id on each build.