SCP: Explicitly put fd 0 into non blocking mode when running simulation on Linux
This change is not actually necessary on Linux, but a bug in Windows
Subsystem for Linux (WSL) doesn't honor tcsetattr() with ICANON disabled
and VMIN and VTIME set to 0 should return immediately when no input data
is available, but it doesn't.  Setting the fd to non blocking produces
equivalent behavior.