SCP: Add "HELP DEVICE devnam"
This allows for device help to be displayed for devices which may have
names that are ambiguous with respect to commands.  For example
if a device named CON exists, HELP CON will display info about the
CONTINUE command rather than the device.  HELP DEVICE CON will
now display that device's HELP info.