cmake build system
CMake-based build system that parallels the makefile-based supported
build system. It does not supercede, replace or otherwise eliminate
the makefile-based build system. It is merely an alternative means to
build simh with a slightly different approach to package dependencies on
Windows (and other platforms, as necessary.)

Documentation in README-cmake.md