makefile: Rename primary target name for vax simulator
make vax the primary build target and microvax3900 the copied result.

The microvax3900 binary name was added when all the different VAX
simulator models were add to simh and each was appropriately named
based on the system model name of each respective simulator.

The vax simulator name was the earliest vax implemented so a binary
of that name is produced for upward compatibility for existing users.
Producing the vax simulator and then copying it to microvax3900
produces equivalent results, but is easier to auto convert to cmake
build configuration files which then will generate a vax project
file when cmake is generating those.