ETHER: Avoid potential compiler warning about version string overrun
gcc 9.3.0 on Ubuntu 20.04 somehow knows that the length of the
string pcap_lib_version() returns can be 256 bytes long.  It then
generates a warning about truncation potential.  The problematic
code path will only be executed on Windows with an old version of
Npcap, but the compiler can't know that.  Make the local buffer
larger to silence the resulting noise.