TMXR: Change output rate limit to 1 ms when cpu is stopped
As reported in #865

When instruction execution is stopped, forced mux output rate limiting
can't leverage instruction execution rate to time inter-character delays.
A direct sleep is used but that has a minimal granularity measured in
milliseconds.  The prior default for a non speed controlled line was 10ms
sleep between character output.  This has been changed to 1ms which
is reasonable for most situations.  Meanwhile, since the remote console
facility uses TMXR, and it has the reasonable ability to generate output
when instructions are not being executed, this is better but not perfect.