MicroVAX2000, VAXStation2000: Properly describe 16MB of memory in CFGTST reg
The CFGTST register MTYPE subfield should describe the additional memory
beyond 2MB on the processor board.  Previous logic attempted to describe
the total system memory and the net result didn't fit into the 3 bit field and
thus said no additional memory is present.

The consequence of this new amount of memory is that ALL of it is tested
during the power on self test and thus it takes significantly longer to get to
the >>> prompt.

As reported in #944