makefile: allow .git to be a file or directory to handle submodules
Change test for -e (exists) instead of -d (directory) so that the .git
entry is detected correctly and git version stamp logic runs as in any
other git repo. The .git entry is a file when the repo is a submodules
and possibly other obscure cases like multiple worktrees.

fixes #988