H316: Fix reading messages from the host interface.
The 1822 IMP-host interface carefully handshakes every bit
transferred.  Thus the DMC facility can read messages piecewise.  The
IMP uses this to read a 32-bit leader first, and process the rest or
the message in another round.

The IMP expects the 1822 end of message bit not to be set until after
the entire message has been read.