doc: Revise the SIMH User's guide to describe startup config files executed
The key change is the addition of:

When a simulator starts execution, the following sequence of simh command
files are executed if they are found:
    1. If a file named simh.ini is located in your HOME directory, it is
        executed.
    2. If the simh.ini file in your HOME directory isn’t found, a file named
        simh.ini in your current working directory is executed if it exists.
    3. If the simulator is invoked with any arguments, then the arguments are
        presumed to be a command file and possible arguments to that command
        file which is executed.
    4. If the simulator is invoked without any arguments, then a command file
        with the same name as the simulator binary with .ini appended that is
        located in the current working directory is executed.

Note, that up to 2 separate command files may be executed on simulator
startup.  The simh.ini file allows the user to define local user
preferences that align with their personal goals for simulator execution
across all simulators that may be used on their system.

Steps 3 and 4 were inherited from simh v3.x.  Steps 1 and 2 was inspired
by conversations with J. David Bryan in April of 2012.  Some how it
never got documented.