PDP11, VAX: Add RF30, RF31, RF71 and RF72 disk types to the RQ simulation
These are added based on simple drive geometry and size information
without any parameters for bad block management (not used in simulation
anyway).
Additionally, RF drives connected to Qbus systems via a KFQSA.  When
connected this way, EACH drive on the DSSI bus appeared to the attached
Qbus system as a separate MSCP controller to the operating system.  This
change does not model that connectivity.