PDP11: Revert DZ device to prior version.
The code in 3.12-2 is only part of a complete solution and thus merely
hangs the device/processor.

The reverted version wouldn't normally have the problem since the
normally partial code was trying to fix since I/O completion happens
right in dz_wr().  However, the "complete it immediately" model could
logically lose data if the sending side were driven harder than the TMXR
TCP transport was able to deliver data to the TCP session.  The easiest
answer is to revert to the earlier v3 DZ code since at least that worked
almost all of the time.

The right answer would be to only signal the transmit completion
interrupt if the transmit didn't end up getting buffered and if it did
get buffered (and not successfully flushed), then schedule a separate
unit to retry the pending output until it actually made it to the TCP
wire.  That separate unit would then signal the interrupt for the line
indicated in the CSR and mark CSR_TRDY.

This is more change to realize relatively little value so simply revert
and be done.