fixup! 3b2: Fix for clock drift when idling
Forgot to attribute this commit to Seth Morabito <web@loomcom.com>
as author.

When the 3B2 simulator was set to allow idling, there was significant
clock drift related to the primary timer unit. It turns out that the
simulator was using `AIO_SET_INTERRUPT_LATENCY` and `sim_rtcn_tick_ack`
incorrectly. They are not needed with the structure of system timers in
the 3B2 architecture.